# packages/models/src/sample-item/sample-item.write.schema.ts
import { z } from 'zod'

import { fields } from '../common/fields.js'

export const SampleItemWriteSchema = z.object({
  name: fields.optionalNullableString,
})

export const SampleItemCreateRequestSchema = SampleItemWriteSchema